workers/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared utilities for Unstract workers"
authors = [
    {name = "Unstract Team", email = "support@unstract.com"},
]
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.3.0,<6.0.0",
    "kombu>=5.3.0,<6.0.0",
    "requests>=2.31.0,<3.0.0",
    "urllib3>=1.26.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "structlog>=23.1.0,<24.0.0",
    "prometheus-client>=0.17.0,<1.0.0",
    "psutil>=5.9.0,<6.0.0",
    "flask>=2.3.0,<3.0.0",
    "werkzeug>=2.3.0,<3.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "pybreaker>=0.8.0,<1.0.0",
    "orjson>=3.9.0,<4.0.0",
    "python-dateutil>=2.8.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "sqlalchemy>=1.4.0,<2.0.0",
    "httpx>=0.27.0,<0.28.0",
    "redis>=4.0.0,<6.0.0",
    "unstract-sdk[azure]>=0.73.1",
    "unstract-connectors",
    "unstract-core",
    "unstract-flags",
    "unstract-tool-registry",
    "unstract-tool-sandbox",
    "unstract-workflow-execution",
    "unstract-filesystem",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
unstract-filesystem = { path = "../../unstract/filesystem", editable = true }
unstract-workflow-execution = { path = "../../unstract/workflow-execution", editable = true }
unstract-tool-sandbox = { path = "../../unstract/tool-sandbox", editable = true }
unstract-tool-registry = { path = "../../unstract/tool-registry", editable = true }
unstract-flags = { path = "../../unstract/flags", editable = true }
unstract-core = { path = "../../unstract/core", editable = true }
unstract-connectors = { path = "../../unstract/connectors", editable = true }
